Modularize args and output generation
